fixed mobs stopping just before players

+++ arena-api/src/main/java/io/github/zap/arenaapi/pathfind/PathOperation.java
@@ -126,14 +126,14 @@ public interface PathOperation {
     }
 
     static PathOperation forEntityWalking(@NotNull Entity entity, @NotNull Set<? extends PathDestination> destinations,
-                                          ChunkCoordinateProvider searchArea, int targetDeviation) {
+                                          ChunkCoordinateProvider searchArea, double targetDeviation) {
         return forAgent(PathAgent.fromEntity(entity), destinations, HeuristicCalculator.DISTANCE_ONLY,
                 SuccessCondition.whenWithin(targetDeviation * targetDeviation),
                 new DefaultWalkNodeExplorer(AversionCalculator.DEFAULT_WALK), DestinationSelector.CLOSEST, searchArea);
     }
 
     static PathOperation forEntityWalking(@NotNull Entity entity, @NotNull Set<? extends PathDestination> destinations,
-                                          int loadRadius, int targetDeviation) {
+                                          int loadRadius, double targetDeviation) {
         return forAgent(PathAgent.fromEntity(entity), destinations, HeuristicCalculator.DISTANCE_ONLY,
                 SuccessCondition.whenWithin(targetDeviation * targetDeviation),
                 new DefaultWalkNodeExplorer(AversionCalculator.DEFAULT_WALK), DestinationSelector.CLOSEST,
