+++ zombies/src/main/java/io/github/zap/zombies/game/mob/goal/OptimizedMeleeAttack.java
@@ -12,7 +12,7 @@ public class OptimizedMeleeAttack extends RetargetingPathfinder {
     private int attackTimer;
 
     public OptimizedMeleeAttack(AbstractEntity entity, AttributeValue[] attributes, double speed, int attackInterval,
-                                float attackReach, int targetDeviation) {
+                                float attackReach, double targetDeviation) {
         super(entity, attributes, speed, targetDeviation);
         this.attackInterval = attackInterval;
         this.attackReach = attackReach;
